Transfer favorite recipes view to domUpdate file
