+++ src/scripts.js
@@ -1,5 +1,14 @@
-// import apiCalls from './apiCalls';
 import './styles.css';
+// import apiCalls from './apiCalls';
+import { fetchResponse } from './apiCalls';
+// domUpdates can I import all functions on one line??
+import { getRecipeBox } from './domUpdates';
+import { searchErrorMsg } from './domUpdates';
+import { clearView } from './domUpdates';
+import { recipeDetails} from './domUpdates';
+import { iconToFull } from './domUpdates';
+import { iconToEmpty } from './domUpdates';
+
 import { RecipeRepository } from '../src/classes/RecipeRepository';
 import { User } from './classes/User';
 import { Recipe } from '../src/classes/Recipe';
@@ -9,7 +18,6 @@ import './images/full-star.png';
 import './images/empty-star.png';
 import './images/full-to-cook.png';
 import './images/empty-to-cook.png';
-import { fetchResponse } from './apiCalls';
 
 // Global Variables
 let allIngredientsData = [];
@@ -193,20 +201,6 @@ function loadTagSearchView() {
   showElement(findByTagView);
 }
 
-function iconToFull(iconName) {
-  let icon = document.querySelector(`.${iconName}-icon`);
-  icon.src = `./images/full-${iconName}.png`;
-  icon.classList.remove(`empty-${iconName}`);
-  icon.classList.add(`full-${iconName}`);
-}
-
-function iconToEmpty(iconName) {
-  let icon = document.querySelector(`.${iconName}-icon`);
-  icon.src = `./images/empty-${iconName}.png`;
-  icon.classList.remove(`full-${iconName}`);
-  icon.classList.add(`empty-${iconName}`);
-}
-
 function addFavoriteRecipe(event) {
   iconToFull('star');
   let recipeToAdd = allRecipeData.find(recipe => {
@@ -238,14 +232,9 @@ function removeFavoriteRecipe(event) {
 }
 
 function favoriteCurrentRecipe() {
-  favoriteRecipesView.innerHTML = '';
+  clearView(favoriteRecipesView);
   currentUser.favoriteRecipes.forEach((recipe) => {
-    favoriteRecipesView.innerHTML += `
-      <div class='box recipe-box'>
-          <img class='test' id=${recipe.id} src=${recipe.image} alt='${recipe.name} image' />
-          <h4 class='recipe-name'>${recipe.name}</h4>
-          <h5 class='recipe-tags'>Tags: ${recipe.tags}</h5>
-      </div>`
+    favoriteRecipesView.innerHTML += getRecipeBox(recipe);
   });
 }
 
@@ -267,12 +256,7 @@ function addToCookRecipe(event) {
 function toCookCurrentRecipe() {
   toCookView.innerHTML = `<h3 class="page-title">Recipes for ${currentUser.name} To Cook:</h3>`;
   currentUser.recipesToCook.forEach((recipe) => {
-  toCookView.innerHTML += `
-    <div class='box recipe-box'>
-      <img id=${recipe.id} src=${recipe.image} alt='${recipe.name} image' />
-      <h4 class='recipe-name'>${recipe.name}</h4>
-      <h5 class='recipe-tags'>Tags: ${recipe.tags}</h5>
-    </div>`
+  toCookView.innerHTML += getRecipeBox(recipe);
   });
 }
 
@@ -296,14 +280,9 @@ function loadAllRecipesView() {
   showElement(searchResultView);
   showElement(allRecipeView);
 
-  allRecipeView.innerHTML = '';
+  clearView(allRecipeView);
   allRecipeStorage.recipes.forEach((recipe) => {
-    allRecipeView.innerHTML += `
-      <div class='box recipe-box'>
-        <img class='test-class' id=${recipe.id} src=${recipe.image} alt='${recipe.name} image'/>
-        <h4 class='recipe-name'>${recipe.name}</h4>
-        <h5 class='recipe-tags'>Tags: ${recipe.tags}</h5>
-      </div>`
+    allRecipeView.innerHTML += getRecipeBox(recipe);
   });
 }
 
@@ -325,47 +304,19 @@ function loadRecipeDetailView(event) {
       return recipe.id === currentRecipe.id;
     });
 
-    recipeDetailCard.innerHTML = `
-      <h3 class='recipe-name'>${currentRecipe.name}</h3>
-      <h4>Ingredients</h4>
-      <p>${currentRecipe.getIngredientNames(allIngredientsData)}</p>
-      <h4>Total Cost</h4>
-      <p>$${currentRecipe.getTotalCostInDollars(allIngredientsData)}</p>
-      <h4>Favorite</h4>
-      <div class="favorite-star" id=${currentRecipe.id}>
-        <img class="star-icon empty-star" id=${currentRecipe.id} src="" alt="star icon"/>
-      </div>
-      <h4>To Cook</h4>
-      <div class="to-cook-tool" id=${currentRecipe.id}>
-        <img class="to-cook-icon empty-to-cook" id=${currentRecipe.id} src="" alt="chef tools icon"/>
-      </div>`;
-    instructionsList.innerHTML = '<h3>Instructions</h3>';
-    currentRecipe.instructions.forEach((instruction) => {
-      instructionsList.innerHTML += `
-        <li>${instruction.instruction}</li>`
-    });
-
+    recipeDetails(recipeDetailCard, currentRecipe, instructionsList, allIngredientsData);
     if(isFavorite) {
       iconToFull('star');
     } else {
       iconToEmpty('star');
     }
 
-    if(isWantingToCook) {
+    if (isWantingToCook) {
       iconToFull('to-cook');
     } else {
       iconToEmpty('to-cook');
     }
   }
-}
-
-function getRecipeBox(recipe) {
-  return `
-  <div class='box recipe-box'>
-      <img id=${recipe.id} src=${recipe.image} alt='${recipe.name} image'/>
-      <h4 class='recipe-name'>${recipe.name}</h4>
-      <p class='recipe-tags'>Tags: ${recipe.tags}</p>
-  </div>`
 }
 
 function searchRecipeByName(searchingFor) {
@@ -374,9 +325,9 @@ function searchRecipeByName(searchingFor) {
     return recipe.name.toLowerCase().includes(searchingFor);
   });
   showElement(searchResultsView);
-  searchResultsView.innerHTML = "";
+  clearView(searchResultsView);
   if(!nameResults.length) {
-    searchResultsView.innerHTML = "<h3>Sorry, nothing found! Try another search.</h3>"
+    searchErrorMsg(searchResultsView);
   } else {
     nameResults.forEach((recipe) => {
       searchResultsView.innerHTML += getRecipeBox(recipe);
@@ -390,9 +341,9 @@ function searchFavRecipeByName(searchingFor) {
     return recipe.name.toLowerCase().includes(searchingFor);
   });
   showElement(favoriteRecipesView);
-  favoriteRecipesView.innerHTML = "";
+  clearView(favoriteRecipesView);
   if(!nameResults.length) {
-    favoriteRecipesView.innerHTML = "<h3>Sorry, nothing found! Try another search.</h3>"
+    searchErrorMsg(favoriteRecipesView);
   } else {
     nameResults.forEach((recipe) => {
       favoriteRecipesView.innerHTML += getRecipeBox(recipe);
@@ -413,9 +364,9 @@ function searchFavRecipeByTag(searchingFor) {
   });
   
   showElement(favoriteRecipesView);
-  favoriteRecipesView.innerHTML = '';
+  clearView(favoriteRecipesView);
   if(!tagResultRecipes.length) {
-    favoriteRecipesView.innerHTML = "<h3>Sorry, nothing found! Try another search.</h3>"
+    searchErrorMsg(favoriteRecipesView);
   } else {
     tagResultRecipes.forEach((recipe) => {
       favoriteRecipesView.innerHTML += getRecipeBox(recipe);
@@ -437,9 +388,9 @@ function searchRecipeByTag(searchingFor) {
 
   showElement(searchResultsView);
   if(!tagResultRecipes.length) {
-    searchResultsView.innerHTML = "<h3>Sorry, nothing found! Try another search.</h3>"
+    searchErrorMsg(searchResultsView);
   } else {
-    searchResultsView.innerHTML = '';
+    clearView(searchResultsView);
     tagResultRecipes.forEach((recipe) => {
       searchResultsView.innerHTML += getRecipeBox(recipe);
     });
